deflake test, convert to int

+++ bamboo/tests/core/test_calculations.py
@@ -127,7 +127,7 @@ class TestCalculations(TestCalculator):
                     # downsample and set low precision for time comparison
                     stored = (parse_date_to_unix_time(self.now) -
                               parse_date_to_unix_time(row['submit_date'])) / 10
-                    result = result / 10
+                    result = int(result / 10)
                     places = 0
                 else:
                     stored = np.float64(row[formula])
